Type getTaskParams and move types to a new file
