+++ src/app/modules/item/pages/item-display/item-display.component.ts
@@ -1,10 +1,10 @@
 import { AfterViewInit, Component, ElementRef, Input, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { DomSanitizer, SafeResourceUrl } from '@angular/platform-browser';
 import { ItemData } from '../../services/item-datasource.service';
-import { taskProxyFromIframe, taskUrlWithParameters, TaskListener, Task, TaskParams, TaskParamsValue }
-  from 'src/app/modules/item/task-communication/task-proxy';
+import { taskProxyFromIframe, taskUrlWithParameters, TaskListener, Task, } from 'src/app/modules/item/task-communication/task-proxy';
 import { EMPTY, forkJoin, interval, Observable, of, Subscription } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
+import { TaskParams, TaskParamsKeyDefault, TaskParamsValue } from '../../task-communication/types';
 
 interface TaskTab {
   name: string
@@ -173,7 +173,7 @@ export class ItemDisplayPlatform extends TaskListener {
     return EMPTY;
   }
 
-  getTaskParams(_keydefault?: [string, TaskParamsValue]) : Observable<TaskParamsValue> {
+  getTaskParams(_keyDefault?: TaskParamsKeyDefault) : Observable<TaskParamsValue> {
     return of(this.taskParams);
   }
 }
